Add mlbench.context module stub
For collecting context around a benchmark.

Context is defined as a key-value pair (the key being a string) giving
a piece of information relevant to the benchmark. A function that returns
such a pair, or a sequence of pairs, is called a _context provider_.

In an upcoming commit, context providers will be added to the benchmark
runner abstract class, where they can either be supplied on construction,
or registered through an entry point (optional).