Add utils and runner discovery test coverage
Utils currently only contain parametrized tests for `ismodule` and
`modulename`.

The runner discovery test gets a nested directory without an `__init__.py`,
which tests that module import also works without a proper package
structure.

What we should check is if we can fast-track the module import if it's
part of an installed package.