Add context provider abstraction, first examples
A context provider is defined as a function taking no arguments and
producing either a key-value pair or a sequence of key-value pairs.

Adds basic context providers for system information from `platform`.
Other info like CPU/GPU config need optional dependencies.

Make the result type a TypedDict consisting of a "context" key and a
"benchmarks" key, holding an array of the benchmark results.

The context can be specified in any benchmark run, and is (as of now)
not a part of the runner state.