Properly type and document the core decorators
Previously, the decorator typing was suffering since no proper overloads
were available for the parentheses vs. no parentheses cases. These are
now added.

As a consequence, both varieties for both decorators now result in
sensible return types. This also paves the way for an easier implementation
of the cartesian product decorator coming up.

As a consequence of the removed params slot, the function in a
parametrized benchmark is replaced by an update-wrapped partial, with
the keyword arguments from the parametrization applied.