Add cartesian product decorator for benchmarks
Adds the third core decorator for parametrizing benchmarks, which is
the cartesian product of variable collections. It is exposed as
`nnbench.product`.

Compared to the "inner product" approach of `nnbench.parametrize`, which
consumes an iterable of dictionaries, this decorator consumes a keyworded
collection of iterables, which it then binds to the benchmark function.
This results in an "outer product" with a number of benchmarks that is
equal to the products of the numbers of values in each iterable.

However, since `functools.partial` binds parameters lazily, users do not
immediately get feedback on whether they supplied the correct argument
names in the iterable variable keyword arguments. This needs to be
improved, by either implementing a check inside the decorator or by
typing it to the decorated benchmark function's param spec (Python 3.11+).