Handle untyped interfaces in `_check` (#18)
Guard match checks of types in function signature and supplied via the params `kwarg` in the benchmark decorators `_check` method against untyped function signatures.

If untyped, we do not check for matching types. We also log a warning to the `stdout` of the logger indicating the untyped parameter and function name.

This implementation also throws a `TypeError` (non-unique types)  when a parameter appears typed and not explicitly typed. Even if the type implicitly matches. 

---------

Co-authored-by: Nicholas Junge <n.junge@appliedai.de>