Refactor and improve combined interface check
Uses the newly defined benchmark functions's interface everywhere.

Fixes a type check which would fail if a variable has two non-identical
types in two different benchmarks, but which are compatible (i.e. one is
a subtype of the other).

Also fixes an error where a missing default parameter in one benchmark
could be overridden by a present one in another benchmark. Now, a missing
default value means a missing default in the resulting union interface.