Refactor type check to be compatible with generics and type hints
For typing types and generics, `issubclass` throws errors. Thus, we
switch to a more stable check that canonicalizes inputs first (i.e.,
calls `typing.get_origin` on its inputs to strip generic parametrization)
and then checks `issubclass` only if the LHS type is a class.

This does not take into account generic parameters, so the user has to
rely on type hints in his IDE to make things work, but we deem this
acceptable until someone complains.