Add namegen argument to `@nnbench.{parametrize,product}` (#51)
* Add namegen argument to `@nnbench.{parametrize,product}`

Switches out the generated name for a custom name, with the default
staying the same as before. The default (function name plus underscore-
joined list of keyword arguments) was refactored into a module-private
name generator.

* Add warning when encountering a benchmark name collision

Contains an explanation as to why the warning is shown. Since we cannot
be sure whether it is a param collision or a faulty namegen, we just
say "Perhaps".