Inline console reporter into `BenchmarkReporter.display()`
Since the ability to print to console is so nice, we change the console
reporter to be available on every reporter descendant by adding it as
the `display()` method on the base.

This means that a) the console reporter file can go, and b) tabulate
has to be upgraded to an obligatory dependency (at least until we have
a better tabular formatter in rich etc.).