Remove console reporter file, add tabulate to project requirements
Also updates the lock files, although not much should change because it
was a dev dependency already.

Removes the console reporter from the registry, which is unused at this
point, and might be ripe for removal until we add a CLI running facility.

Updates all the docs to use the base `BenchmarkReporter()` and its
`display()` method.