Add a basic file reporter
Implements a polymorphic read/write facility via a simple backend
abstraction. A backend here is a pair of serialization and deserialization
callbacks with a strict interface.

Starting backends are JSON, CSV (untested), and YAML (untested).

Currently, the choice of inlining context vs. saving only the two top
level keys is not given, which results in errors when trying to read
records with duckDB. This will be addressed in a follow-up.