Add barebones duckDB reporter
This is a bit of a misnomer, since the class is currently just a file
reporter with some extra filtering facilities. To improve, we need to
open up a SQL query builder facility to actually obtain analytics from
the database without too much effort.

In that light, it may be beneficial to outsource this part of the process
if supporting all relevant SQL features turns out to be too cumbersome.