Add dict roundtrip methods to `BenchmarkRecord`, improve file drivers
This is a more comprehensive take on context handling before/after a
serialization. Options are now omitting, inlining, or flattening the
context.

Changes the `BenchmarkRecord` class to a dataclass, meaning everything
gets switched over to dotted access.

Turns the file reporter into an IO mixin, which is a good building block
for subsequent uses in other reporters.