Unify batching in file drivers, remove read/write stubs on base reporter
Since the reporter interface is so dynamic (i.e., file I/O needs vastly
different arguments than DB I/O), we scrap the read/write APIs from the
base class and add it as needed in the child classes.

Adds batched (i.e., multi-record) read/write support and a `FileReporter`
class that is just the sum of the `FileIO` and `BenchmarkReporter`
interfaces.