Add compression algorithms, selection in FileIO sink
Works similarly to the file driver lookup, except the compression can
also be None, in which case a normal file descriptor is used.

Canonicalizes the extension of the resulting file regardless of which
file name is used based on driver and compression algorithm. This might
be surprising behavior, but simplifies driver and compression inference
significantly on roundtrips.

Scraps the custom registries in FileIO constructor, because we expose
reg/dereg hooks for file drivers and compressions.