Make `contextkeys()` an API to handle separator properly
The BQ example revealed that we cannot just store the keys as a property
on our class, since they might not be digestible by a database if certain
illegal characters are used.

For BQ, this was the case with the default "." separator, which means we
have to pipe the separator through in the `BenchmarkRecord.compact()`
call. Hence, the context keys need to be fetched dynamically.