Add file IO roundtrip tests, tweak APIs to expected test results
Most of these tests passed out of the box, except for CSV, which incurred
some trouble with flattened contexts and quotechars.

As a consequence, the CSV reader and writer get a small block of control
flow, which JSON-loads the context and _contextkeys attributes if present,
since Python-native CSV coerces everything to string. This requires some
quotechar manipulation, and changing over the `_contextkeys` attribute
to a list of strings instead of a tuple, since json.load cannot deal with
round parentheses.