Add links to the quickstart and example docs
The presented lightning demo is very lean and concise, but not really
illustrative of how the project should be used.

For further reading, we add links to the quickstart, which showcases a
scikit-learn model, and to the examples, which give a variety of examples
for different applications.

Also adds the new(ish) examples guides to the index page as links with
a more concise content summary each.