Implement key collision logic in context
The `.add` and `.update` methods of the `Context` now check for
key collisions. With the `replace=False` `kwarg` the user can cus-
tomize the behaviour. Upon collision a ValueError is raised.

Remove collision logic from `runner.run` method as it is now hand-
led in the `Context.add` method.

The `.add` method is implemented in term of `.update` to stay DRY.