Migrate tests to new params assumptions
Partial parametrizations are not bound eagerly to the benchmark functions
anymore, which makes it simpler to inject memos and de-memoize variables
just in time for execution.

What is left is validation that a subsequent benchmark of models with
intermittent garbage collection actually reaps each model after the
benchmark is done.