Inject state class into setUp and tearDown tasks
in the `runner.run()` method. This is necessary such that the setUp and
tearDown tasks know the benchmark states. Namely, how many bench-
marks are in the benchmark family and the index of the current.

In a follow up implementation of a cache we will use the index
and family length to compute a condition to empty the cache.