(wip) Add caching on type-level
with the help of two classes. The MemoCache is not intended to be
instantiated but holds a global cache. One can interact with this
cache via classmethods.

The `Memo` class is to be used as a decorator on whatever one wants
to cache. It computes a value lazily and caches it within
`MemoCache`.

You can delete the class to empty the cache.