Add global memo cache and integrate  (#130)
Add a global cache dict in `nnbench.types`. Users subclass the 
`Memo` class with the `@cached_memo` decorator applied
to the override of `__call__` method. The call should return the
object that should be cached. 

`del` on the subclassed `Memo` also removes the wrapped 
item from the cache. 

Add utility methods for cache management as well as tests.

---------

Co-authored-by: Nicholas Junge <n.junge@appliedai.de>