Change guide to HuggingFace model benchmarking (#133)
Now advertises memos as the canonical way of using caching+lazy loading
in combination for memory efficiency.

Changes the example structure to three standalone Python files instead
of a project. This is because we disentangled the files by removing the
argument parser, and trivialized the runner.py file.

Changes the guide name to mention HuggingFace as the tech of interest in
this example.