refactor: Remove context class, back to raw dictionaries
With flattening and unflattening as functional methods moved into
`nnbench.util`.

The class was pretty much just a wrapper around a dict with nesting
support on a separator, and didn't provide a lot of value on its own.

Since we already typed the context provider as a callable returning a raw
dictionary, it doesn't create any churn to switch the context out for a
raw dict, either.