Change benchmark record export capabilities, update file IO
It is useful to be able to export a record to both a struct (JSON, YAML)
and to a list of results (tabular formats like CSV, Parquet, databases).

The previous way of choosing either mode via a string literal was not
great, so now, we restructure the API to `BenchmarkRecord.to_{json,list}`
for either export, and make BenchmarkRecord.expand() capable of dealing
with either.