Add `nnbench compare` subcommand, comparison function
Works using metric (benchmark) names as columns, and records as lines.
Alternatively, we could support metrics as rows and record IDs as columns
in the future.

Contains logic to add parameter values and context values to the table
as columns, also directly on the command line. In the future, we could
make a comparison class that can be subclassed to avoid option clutter on
the command line for customizing the rich table display.