Some tweaks related to context ingestion
Unfortunately, the first prototype for builtin context provider support
from the command line was not successful.

The alternative is to provide values as
key=`from nnbench.context import xy; print(xy()())`, but that returns the
serialized JSON struct, which is not all that useful. Maybe we can get
around that by optionally processing JSON-like values into the context.