Add a fixture system to nnbench
Fixtures give a CLI user the option of defining benchmark input parameters
as "fixtures", i.e. results of functions locate in a special named module,
enabling better customization and configuration.

Unlike pytest, we operate only on raw Python callables, and do not expose
features like scopes, fixture names, and IDs. This is because our
motivation is slightly different from that of pytest, in that we need to
track parameter values and try to keep them constant ("session-scoped")
across benchmarks, except in niche cases.

Our magic filename is `conf.py`, similarly to conftest.py for pytest.

This change incurs changes in how we construct and typecheck parameters -
since we now allow heterogeneous parameters in benchmarks by virtue of
fixtures, we need to typecheck each individual parameter set. This will
be done in a followup.