Add nnbench CLI guide (#179)
* Print help and exit 1 on no args, add benchmark run name to tabular comparison

This makes it easier to see which records the rows come from.

* Add nnbench CLI guide

Contains all commands, their usages and use scenarios, and an appendix section
on how to configure the CLI experience via `pyproject.toml`.

* Move CLI docs into their own subdirectory, split off pyproject.toml doc

This is the most sensible configuration I can think of. We can decide if
we want to restructure the documents currently under "User Guide" into a
section such as "Python SDK".