Remove `nnbench.BenchmarkRunner` in favor of `nnbench.collect()` and `nnbench.run()`
The class had little to no useful state, anyways, and these two standalone APIs play
way better with threading and concurrency, so we opt for functional APIs instead of
a monolithic class.

This is effectively an idiom change - before, we were instantiating as
`runner = nnbench.BenchmarkRunner`, whereas now, we go the extra mile to
collect in the open (`benchmarks = nnbench.collect(path, tags=...)`), and
then pass them as the first argument to the "new" `nnbench.run()` API, which
is just the BenchmarkRunner.run() with the path and tags arguments removed.

This also prevents erroneous caching and persistence of large params to an extent,
since the benchmark list is made explicit.