cli: Add support for parallelism via `multiprocessing`
This allows process parallelism for execution of benchmarks. It works such
that discovered files are passed to `Pool.map()`, which prompts each process
to independently discover benchmarks in files, loading the necessary fixtures,
and executing.

Resulting records are then merged on benchmarks only, since the context is (for now)
assumed equal across benchmarks.