feat(cli): Add `-j` option for running benchmarks in parallel (#206)
* feat(cli): Add `-j` option for running benchmarks in parallel

Usage: `nnbench run <benchmarks> -jN`.

Similar to make, but using process parallelism. The default is to not use
multiple processes. What's left to decide is whether to require joblib as
a hard dependency for parallelism, to simplify the control flow in the CLI.

* cli: Add support for parallelism via `multiprocessing`

This allows process parallelism for execution of benchmarks. It works such
that discovered files are passed to `Pool.map()`, which prompts each process
to independently discover benchmarks in files, loading the necessary fixtures,
and executing.

Resulting records are then merged on benchmarks only, since the context is (for now)
assumed equal across benchmarks.

* tests: Add CLI parallel execution test

Calls `nnbench run benchmarks -j2`, testing that execution of two trivial
benchmarks with a 10 second sleep finishes faster than the sequential 20 seconds.