cli: Refactor reporter portion to use multiple dispatch
Based on the value of outfile, i.e. the value of the `-o` switch.

The introduced taxonomy is stdout | file | database | service, and loads
a different IO class depending on the detected type of IO requested.

For files, inferring the IO type is a bit harder, since local files are
usually given without a leading protocol. For this case, we look at the
file extension as well.

Every other IO is detected by the leading protocol of the given URI, so
`mlflow://experiment/run/subrun` leads to the MLflow IO being requested.