cli: Add --jsonifier switch to support user-given params JSONification
This is better than the proposed context provider refactoring. What's missing
is the implementation as a TOML value either in [tool.nnbench] or
[tool.nnbench.run].

NB: For importing a resource outside of an installed project (I attempted this
manually), we have to insert something into sys.path or tweak PYTHONPATH.
It would be good to do this automatically, such that the user does not have to
fiddle with PYTHONPATH themselves.