feat: Add comparison class (#223)
* feat: Add comparison class

The necessary building blocks are a comparison operator enum containing
all comparison operators, and a comparison maker.

* compare: Add a comparison class, example implementation

The comparison will become an abstract base class, this is currently only
a proof of concept to pitch to users.

* comparison: Add `successful` property, extend to >2 records

Right now, only the "one vs. all" comparison mode is supported.

* cli: Remove extra context and parameter groups for now

* comparisons: Add abstract base class, comparators, JSON loading for CLI

To load a comparison config (JSON) on the CLI, pass `--comparison-file=$FILE`,
where the file points to a JSON containing key-value pairs of metric names and
comparators.

Fixes the table renderer to inherit from an abstract base, make comparisons
optional for metrics, and skip comparisons for values which are missing in
other records.