chore: Rename BenchmarkRecord -> BenchmarkResult, add `(to|from)_records` API (#224)
This was an unfortunate misnomer for us, since we want to stream benchmark results
(i.e. metrics) to and from databases too, and in that context, the "record" nomenclature
is more fitting for a single row (i.e., a single benchmark output).

Hence, do the aforementioned rename to avoid ambiguity.

This also adds a result-level `timestamp` attribute, which supersedes the per-benchmark
timestamp, which was unused.