feat: Add SQLite reporter
Responsible for reading and writing results to SQLite databases.

This requires a tweak to `BenchmarkResult.from_records()`, because in principle,
an SQLite query result can contain an arbitrary number of runs (and so can files
for that matter). Thus, the return type is changed to yield a list of benchmarks.

Usage: nnbench run benchmarks.py -o sqlite://my-database.db.