fix: Rename output path argument to "path", fix MLflow reporter
This fixes some fallout of the removal of iterable results from the reporter interface.

It also renames the input/output path argument to "path". Naming them differently for
files (where it was "file" or "fp") and databases (where it was "uri") is technically a
Liskov violation, if the arguments are not marked as positional only.